Changed contructor-arg injection to property injection changed the base save and delete method

+++ com/service/BaseService.cfc
@@ -58,21 +58,30 @@ component displayname="Base Service" persistent="false" accessors="true" output=
 		}
 	}
 	
-	public void function delete(required any entity){
-		getDAO().delete(entity=arguments.entity);
+	public boolean function delete(required any entity){
+		var deleted = false;
+		if(!arguments.entity.hasErrors()) {
+			getDAO().delete(entity=arguments.entity);
+			deleted = true;
+		} else {
+			transactionRollback();
+		}
+		return deleted;
 	}
 	
-	public any function save(required any entity) {;
+	public any function save(required any entity, struct data) {
+		if(structKeyExists(arguments,"data")){
+			arguments.entity.populate(arguments.data);
+		}
         getValidator().validateObject(entity=arguments.entity);
 		
 		if(!arguments.entity.hasErrors()) {
 			arguments.entity = getDAO().save(entity=arguments.entity);
 		} else {
 			transactionRollback();
-			trace( text="rolled back save within base service");
+			//trace( text="rolled back save within base service");
 		}
-		
 		return arguments.entity;
-	}
-	
+	}	
+
 }
